Move helper functions out into a library
